Tests/LibWeb/Text/expected/indexeddb-empty-transaction-autocommit.txt

tx created
tx.complete
DONE
